v1.1

+++ db.php
@@ -1,36 +1,32 @@
 <?php
-// MYSQL database functions [db.php] made by ekk v1.0
+// MYSQL database functions [db.php] made by ekk v1.1
 
 	// include  database config
-	include 'modules/config.php';
-
-	// creating database link
-	$dbLink = Null;
-
+	include 'config.php';
 	
 	// -- [creating connection to database] -- //
 	function dbConnect() {
-		global $dbBaseName, $dbUserName, $dbPassword, $dbHostName, $dbLink;
+		global $dbBaseName, $dbUserName, $dbPassword, $dbHostName;
 		$dbLink = mysqli_connect($dbHostName, $dbUserName, $dbPassword, $dbBaseName);
 		mysqli_query($dbLink, "SET NAMES 'utf8';");
 		mysqli_query($dbLink, "SET CHARACTER SET 'utf8';");
 		mysqli_query($dbLink, "SET SESSION collation_connection = 'utf8_general_ci';");
+		return $dbLink;
+		
 	}
 
 	
 	// -- [close connection to database] -- //
-	function dbDisconnect() {
-		global $dbLink;
+	function dbDisconnect($dbLink) {
 		mysqli_close($dbLink);
 	}
 
 	
 	// -- [make query] -- //
 	function dbQuery($query) {
-		global $dbLink;
-		dbConnect();
+		$dbLink = dbConnect();
 		$ret =  mysqli_query($dbLink, $query);
-		dbDisconnect();
+		dbDisconnect($dbLink);
 		return $ret;
 	}
 	
@@ -95,26 +91,26 @@
 			if($c != count($values)) $q .= ", ";
 		}
 		$q .= ")";
-		dbQuery($q);
+		return dbQuery($q);
 	}
 	
 
 	// -- [delete line] -- //
 	function dbDeleteLine($table, $column, $value) {
+		if(!dbGetLine($table, $column, $value)) return false; 
 		$q = "DELETE FROM `".$table."` WHERE `".$column."` = '".$value."'";
-		dbQuery($q);
+		return dbQuery($q);
 	}
 
-	
 	// -- [update cells by column = value] -- //
-	function dbUpdateCells($table, $column, $value, $columns, $values) {
+	function dbUpdateCells($table, $column, $value, $values) {
 		$q = "UPDATE `".$table."` SET ";
-		for($i=0; $i < count($columns); $i++) {
-			$q .= "`".$columns[$i]."` = '".$values[$i]."'";
-			if($i != count($columns)-1) $q .= ", ";
+		foreach ($values as $col => $val) {
+			$q .= "`".$col."` = '".$val."'";
+			if(next($values)) $q .= ", ";
 		}
 		$q .= " WHERE `".$column."` = '".$value."'";
-		dbQuery($q);
+		return dbQuery($q);
 	}
